fix(cli): inline upgrade banner into main.go for Windows single-file build
The Windows CI step builds with `go build ./main.go` (a single file,
not the package), which excluded the sibling `oss_banner.go` and
caused `undefined: printEnterpriseUpgradeBanner` on Windows.

Moving the function into main.go fixes build-windows-amd64 without
changing CI. Behavior is identical to the prior version — same dedup
ordering (post sudo-reexec gate in start(), agent-name skip,
KEPLOY_INDOCKER skip), same stderr-only output, same NO_COLOR and
--disable-ansi handling.
